ci: Fix the Dockerfile and build from CI
Fix the Dockerfile and add the Docker build to the CI.
The docker build error was:
```
1.359 E: Failed to fetch http://security.debian.org/debian-security/dists/stretch/updates/main/binary-amd64/Packages  404  Not Found
1.359 E: Failed to fetch http://deb.debian.org/debian/dists/stretch/main/binary-amd64/Packages  404  Not Found
1.359 E: Failed to fetch http://deb.debian.org/debian/dists/stretch-updates/main/binary-amd64/Packages  404  Not Found
1.359 E: Some index files failed to download. They have been ignored, or old ones used instead.
```

Also use a minimal image for the final build stage.
New image size is 240MB vs 4.4G for the old one.

In a follow up PR we could add multi-arch build support and automatic
image upload to Docker registries.